refactor: 질문 전송 스케줄러 분리
